Upright MLDB computation seems inconsistent with regular
In the computation of upright MLDB descriptors, it looks like the idea is to assume the angle is zero. In the computation of `rrx` and `rry` in the rotated MLDB, though, this would imply `rrx=ry`. This may not affect performance, but it would be nice to understand what is going on here. If the intention is for Upright to be the same as the other one with a zero angle, this looks like a bug either here or in the regular MLDB.